feat(MigrateDesign): Edit & Delete Registry
Edit & Delete Registry are working but, using old design
