+++ ClientApp/src/app/components/indicator-detail/indicator-detail-registry/indicator-detail-registry.component.ts
@@ -1,9 +1,14 @@
-import { Component, OnInit, Input, ViewEncapsulation } from '@angular/core';
+import { Component, OnInit, Input, ViewEncapsulation, TemplateRef } from '@angular/core';
 import { Registry } from '../../../shared/models/registry';
 import { Document } from '../../../shared/models/document';
 import { RegistryType } from '../../../shared/models/registryType';
 import { RegistryService } from '../../../services/registry/registry.service';
+import { IndicatorService } from '../../../services/indicator/indicator.service';
 import { NgbPanelChangeEvent } from '@ng-bootstrap/ng-bootstrap';
+
+// Ngx-Bootstrap
+import { BsModalService } from 'ngx-bootstrap/modal';
+import { BsModalRef } from 'ngx-bootstrap/modal/bs-modal-ref.service';
 
 @Component({
   selector: 'app-indicator-detail-registry',
@@ -18,13 +23,23 @@ export class IndicatorDetailRegistryComponent implements OnInit {
   @Input()
   public registries: Registry[];
 
-  constructor(private registryService: RegistryService) { }
+
+
+  @Input()
+  public registriesType: number;
+
+  public registry: Registry = null; // For EditRegistry
+  public editModalRef: BsModalRef;
+
+  constructor(private registryService: RegistryService,
+    private indicatorService: IndicatorService,
+    private modalService: BsModalService) { }
 
   ngOnInit() {
   }
 
   private deleteDocument(registry: Registry, document: Document) {
-    const result = confirm('Está seguro que desea elimianr el documento: ' + document.documentName);
+    const result = confirm('Está seguro que desea elimianr el documento: ' + document.name);
     if (registry.documents.length === 1) {
       alert('Debe existir al menos un documento de respaldo para el registro');
       return;
@@ -42,6 +57,29 @@ export class IndicatorDetailRegistryComponent implements OnInit {
       if (index !== -1) {
         registry.documents.splice(index, 1);
       }
+    }
+  }
+
+  openModalEditRegistry(template: TemplateRef<any>, selectedRegistry: Registry) {
+    this.registry = selectedRegistry;
+    this.editModalRef = this.modalService.show(template);
+  }
+
+  private deleteRegistry(registry: Registry) {
+    const date: Date = new Date(registry.date);
+    const formatedDate: string = date.getDate() + '-' + (+date.getMonth() + 1) + '-' + date.getFullYear();
+    const result = confirm('Está seguro que desea eliminar el registro: \n' + formatedDate + ' - ' + registry.name);
+
+    if (result) {
+      this.indicatorService.deleteRegistry(registry).subscribe(
+        data => {
+          const index = this.registries.indexOf(registry);
+          this.registries.splice(index, 1);
+          window.location.reload(true);
+        },
+        err => console.error(err)
+      );
+      
     }
   }
 }
